Remove Expresison Clearing for partial hash compare performance
